+++ example/test/_export_files.dart
@@ -1,0 +1,17 @@
+export 'package:flutter/material.dart';
+
+export 'package:flutter_test/flutter_test.dart';
+
+export 'package:more_extensions/more_extensions.dart';
+
+export 'package:more_extensions/get_utils.dart';
+
+export 'extensions/_export_extensions.dart';
+
+export 'get_utils/string_extension_tests.dart';
+
+export 'get_utils/test_get_utils.dart';
+
+export 'package:example/main.dart';
+
+export 'package:state_set/state_set.dart';