+++ example/test/extensions/_export_extensions.dart
@@ -1,0 +1,22 @@
+/// Copyright 2024 Andrious Solutions Ltd. All rights reserved.
+/// Use of this source code is governed by a 2-clause BSD License.
+/// The main directory contains that LICENSE file.
+///
+///          Created 09 September, 2024
+///
+
+export '_test_extensions.dart';
+
+export 'test_context_extensions.dart';
+
+export 'test_double_extensions.dart';
+
+export 'test_duration_extensions.dart';
+
+export 'test_dynamic_extensions.dart';
+
+export 'test_num_extensions.dart';
+
+export 'test_string_extensions.dart';
+
+export 'test_widget_extensions.dart';