+++ example/test/extensions/test_num_extensions.dart
@@ -5,11 +5,11 @@
 ///          Created 17 March, 2023
 ///
 
-import '../_unit_test_view.dart';
+import '../_export_files.dart';
 
 void testNumExtension() {
   //
-//  test('Test Num Extension', () async {});
+//  test('Test Num Extension', () async {
 
   123.isLowerThan(12);
 
@@ -28,4 +28,5 @@ void testNumExtension() {
   0.3.hours;
 
   .5.days;
+  //});
 }