+++ example/test/extensions/test_widget_extensions.dart
@@ -5,7 +5,7 @@
 ///          Created 17 March, 2023
 ///
 
-import '../_unit_test_view.dart';
+import '../_export_files.dart';
 
 void testWidgetExtension() {
   //