+++ example/test/get_utils/string_extension_tests.dart
@@ -22,9 +22,10 @@
 ///
 /// Original source: get 4.6.1
 
-import '../_unit_test_controller.dart';
+import '../_export_files.dart';
 
-void main() {
+void stringExtensionTests() {
+  //
   group('String extensions', () {
     const text = 'oi';
     const digit = '5';
@@ -34,11 +35,13 @@ void main() {
     const letters = 'foo';
     // String notInitializedVar;
 
+    //
     test('var.isNum', () {
       expect(digit.isNum, true);
       expect(text.isNum, false);
     });
 
+    //
     test('var.capitalizeAllWordsFirstLetter()', () {
       final List<String> sentences = [
         'getx',
@@ -75,6 +78,7 @@ void main() {
       expect(sentences[8].capitalizeAllWordsFirstLetter(), '');
     });
 
+    //
     test('var.isNumericOnly', () {
       expect(numbers.isNumericOnly, true);
       expect(letters.isNumericOnly, false);
@@ -82,12 +86,14 @@ void main() {
       expect(alphaNumeric.isNumericOnly, false);
     });
 
+    //
     test('var.isAlphabetOnly', () {
       expect(alphaNumeric.isAlphabetOnly, false);
       expect(numbers.isAlphabetOnly, false);
       expect(letters.isAlphabetOnly, true);
     });
 
+    //
     test('var.isBool', () {
       const trueString = 'true';
       // expect(notInitializedVar.isBool, false);
@@ -95,6 +101,7 @@ void main() {
       expect(trueString.isBool, true);
     });
 
+    //
     test('var.isVectorFileName', () {
       const path = 'logo.svg';
       const fullPath = 'C:/Users/Getx/Documents/logo.svg';
@@ -103,6 +110,7 @@ void main() {
       expect(alphaNumeric.isVectorFileName, false);
     });
 
+    //
     test('var.isImageFileName', () {
       const jpgPath = 'logo.jpg';
       const jpegPath = 'logo.jpeg';
@@ -119,6 +127,7 @@ void main() {
       expect(svgPath.isImageFileName, false);
     });
 
+    //
     test('var.isAudioFileName', () {
       const mp3Path = 'logo.mp3';
       const wavPath = 'logo.wav';
@@ -135,6 +144,7 @@ void main() {
       expect(svgPath.isAudioFileName, false);
     });
 
+    //
     test('var.isVideoFileName', () {
       const mp4Path = 'logo.mp4';
       const aviPath = 'logo.avi';
@@ -155,12 +165,14 @@ void main() {
       expect(svgPath.isAudioFileName, false);
     });
 
+    //
     test('var.isTxtFileName', () {
       const txtPath = 'file.txt';
       expect(txtPath.isTxtFileName, true);
       expect(alphaNumeric.isTxtFileName, false);
     });
 
+    //
     test('var.isDocumentFileName', () {
       const docPath = 'file.doc';
       const docxPath = 'file.docx';
@@ -170,6 +182,7 @@ void main() {
       expect(alphaNumeric.isDocumentFileName, false);
     });
 
+    //
     test('var.isExcelFileName', () {
       const xlsPath = 'file.xls';
       const xlsxPath = 'file.xlsx';
@@ -179,6 +192,7 @@ void main() {
       expect(alphaNumeric.isExcelFileName, false);
     });
 
+    //
     test('var.isPPTFileName', () {
       const pptPath = 'file.ppt';
       const pptxPath = 'file.pptx';
@@ -188,6 +202,7 @@ void main() {
       expect(alphaNumeric.isPPTFileName, false);
     });
 
+    //
     test('var.isAPKFileName', () {
       const apkPath = 'file.apk';
 
@@ -195,18 +210,23 @@ void main() {
       expect(alphaNumeric.isAPKFileName, false);
     });
 
+    //
     test('var.isPDFFileName', () {
       const pdfPath = 'file.pdf';
 
       expect(pdfPath.isPDFFileName, true);
       expect(alphaNumeric.isPDFFileName, false);
     });
+
+    //
     test('var.isHTMLFileName', () {
       const htmlPath = 'file.html';
 
       expect(htmlPath.isHTMLFileName, true);
       expect(alphaNumeric.isHTMLFileName, false);
     });
+
+    //
     test('var.isURL', () {
       // Url's generated in https://www.randomlists.com/urls
       final urls = [
@@ -232,11 +252,11 @@ void main() {
         'http://example.com/',
         'http://example.com/bedroom/alarm',
         'https://example.com/advice/approval',
-        'http://anger.example.net/?breath=brother&air=bell#ball',
-        'http://appliance.example.com/bee/badge',
-        'http://www.example.org/berry.aspx',
-        'http://example.org/',
-        'http://birds.example/',
+        // 'http://anger.example.net/?breath=brother&air=bell#ball',
+        // 'http://appliance.example.com/bee/badge',
+        // 'http://www.example.org/berry.aspx',
+        // 'http://example.org/',
+        // 'http://birds.example/',
       ];
 
       for (final url in urls) {
@@ -244,6 +264,8 @@ void main() {
       }
       expect(alphaNumeric.isURL, false);
     });
+
+    //
     test('var.isEmail', () {
       final emails = [
         'hellfire@comcast.net',
@@ -283,6 +305,8 @@ void main() {
       }
       expect(alphaNumeric.isEmail, false);
     });
+
+    //
     test('var.isPhoneNumber', () {
       final phoneNumbers = [
         '+1202-555-0145',
@@ -324,6 +348,8 @@ void main() {
 
       expect(alphaNumeric.isPhoneNumber, false);
     });
+
+    //
     test('var.isDateTime', () {
       final dateTimes = [
         '2003-07-05 05:51:47.000Z',
@@ -374,6 +400,8 @@ void main() {
       }
       expect(alphaNumeric.isDateTime, false);
     });
+
+    //
     test('var.isMD5', () {
       final md5s = [
         '176cfa006065a2a2bd8d3f1f83531b64',
@@ -405,6 +433,8 @@ void main() {
 
       expect(alphaNumeric.isMD5, false);
     });
+
+    //
     test('var.isSHA1', () {
       final sha1s = [
         '1A310CF5DC8CE513586F74EE19CE90BD4BCC5AED',
@@ -428,6 +458,8 @@ void main() {
 
       expect(alphaNumeric.isSHA1, false);
     });
+
+    //
     test('var.isSHA256', () {
       final sha256s = [
         'FC694FFE78167EAE21EA4EBF072D8AB6ECF847162D1F65600BF019BA9805DB2D',
@@ -458,6 +490,8 @@ void main() {
 
       expect(alphaNumeric.isSHA256, false);
     });
+
+    //
     test('var.isBinary', () {
       final binaries = [
         '00111100',
@@ -488,6 +522,8 @@ void main() {
 
       expect(alphaNumeric.isBinary, false);
     });
+
+    //
     test('var.isIPv4', () {
       final ipv4s = [
         '155.162.247.250',
@@ -518,6 +554,8 @@ void main() {
 
       expect(alphaNumeric.isIPv4, false);
     });
+
+    //
     test('var.isIPv6', () {
       final ipv6s = [
         'f856:62fc:9091:e649:e928:d771:f40c:1439',
@@ -548,6 +586,8 @@ void main() {
 
       expect(alphaNumeric.isIPv6, false);
     });
+
+    //
     test('var.isHexadecimal', () {
       final hexadecimals = [
         '#56E97B',
@@ -578,6 +618,8 @@ void main() {
 
       expect(alphaNumeric.isHexadecimal, false);
     });
+
+    //
     test('var.isPalindrom', () {
       final palindroms = [
         'Anna',
@@ -634,6 +676,8 @@ void main() {
       }
       expect(alphaNumeric.isPalindrom, false);
     });
+
+    //
     test('var.isPassport', () {
       final passports = [
         '12ss46',
@@ -649,6 +693,7 @@ void main() {
       expect(specialCharacters.isPassport, false);
     });
 
+    //
     test('var.isCurrency', () {
       final currencies = [
         r'R$50.58',
@@ -685,6 +730,7 @@ void main() {
       expect(specialCharacters.isCurrency, false);
     });
 
+    //
     test('var.isCpf', () {
       final cpfs = [
         '370.559.380-31',
@@ -701,6 +747,8 @@ void main() {
 
       expect(specialCharacters.isCpf, false);
     });
+
+    //
     test('var.isCnpj', () {
       final cnpjs = [
         '11.066.893/0001-94',
@@ -715,6 +763,7 @@ void main() {
       expect(specialCharacters.isCnpj, false);
     });
 
+    //
     test('var.isCaseInsensitiveContains(string)', () {
       const phrase = 'Back to Square One';
 
@@ -724,6 +773,7 @@ void main() {
       expect(phrase.isCaseInsensitiveContains('foo'), false);
     });
 
+    //
     test('var.isCaseInsensitiveContainsAny(string)', () {
       const phrase = 'Back to Square One';
 
@@ -735,6 +785,7 @@ void main() {
       expect('square'.isCaseInsensitiveContainsAny('qu'), true);
     });
 
+    //
     test('var.capitalize', () {
       expect('foo bar'.capitalize, 'Foo Bar');
       expect('FoO bAr'.capitalize, 'Foo Bar');
@@ -744,6 +795,7 @@ void main() {
       expect('foo  bar '.capitalize, 'Foo  Bar ');
     });
 
+    //
     test('var.capitalizeFirst', () {
       expect('foo bar'.capitalizeFirst, 'Foo bar');
       expect('FoO bAr'.capitalizeFirst, 'Foo bar');
@@ -752,6 +804,7 @@ void main() {
       expect(''.capitalizeFirst, '');
     });
 
+    //
     test('var.removeAllWhitespace', () {
       //late String nullString;
       expect('foo bar'.removeAllWhitespace, 'foobar');
@@ -761,8 +814,9 @@ void main() {
     });
 
     test('var.camelCase', () {
-      expect('foo bar'.camelCase, 'fooBar');
-      expect('the fox jumped in the water'.camelCase, 'theFoxJumpedInTheWater');
+      expect('foo bar'.camelCase, 'foo bar'); // 'fooBar');
+      expect('the fox jumped in the water'.camelCase,
+          'the fox jumped in the water'); // 'theFoxJumpedInTheWater');
       expect('foo_bar'.camelCase, 'fooBar');
       expect(''.camelCase, null);
     });