+++ example/test/get_utils/test_get_utils.dart
@@ -5,9 +5,10 @@
 ///          Created 17 March, 2023
 ///
 
-import '_unit_test_controller.dart';
+import '../_export_files.dart';
 
 void testGetUtils(WidgetTester tester) {
+  //
   GetUtils.nil(null);
 
   GetUtils.hasCapitalletter('test');