+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 1.1.0
+August 24, 2024
+- Commented out: // if (UniversalPlatform.isDesktop) {
+- Supplied tests and example app
+
 ## 1.0.0
 August 22, 2024
 - Initial commit